Added capability to read commands from a text file.
When other lolbins are used instead of `rundll32.exe`, `GetCommandLineW()` cannot be relied upon.
As a workaround, the program now tries to read and execute the contents of `C:\Window\Tasks\cmd.txt` first.
If no such file is present, commands specified with `-c` would be executed.
If `-c` is not specified either, PowerShell console is launched.